+++ packages/flocks/lib/src/molecules/choice_chip/choice_chip.dart
@@ -1,0 +1,2 @@
+export 'app_choice_chip.dart';
+export 'app_choice_chip_bar.dart';